+++ tests/unit/core/test_wagtail_hooks.py
@@ -200,3 +200,42 @@ def test_estimated_read_time_calculation(rf, domestic_homepage):
         request=request
     )
     assert response > 0
+
+
+@pytest.mark.django_db
+def test_set_lesson_pages_topic_id(rf, topic_with_lessons):
+
+    request = rf.get('/')
+    request.user = AnonymousUser()
+
+    # Rest the topic_block_if for all lessons
+    curated_page = topic_with_lessons[0][0]
+    for topic in topic_with_lessons[0][1]:
+        topic.topic_block_id = None
+        topic.save()
+
+    response = wagtail_hooks.set_lesson_pages_topic_id(
+        page=curated_page,
+        request=request
+    )
+    for topic in response.topics:
+        for lesson in topic.value['pages']:
+            assert lesson.specific.topic_block_id == topic.id
+
+
+@pytest.mark.django_db
+def test_set_lesson_pages_topic_id_already_set(rf, topic_with_lessons):
+
+    request = rf.get('/')
+    request.user = AnonymousUser()
+
+    # Rest the topic_block_if for all lessons
+    curated_page = topic_with_lessons[0][0]
+
+    response = wagtail_hooks.set_lesson_pages_topic_id(
+        page=curated_page,
+        request=request
+    )
+    for topic in response.topics:
+        for lesson in topic.value['pages']:
+            assert lesson.specific.topic_block_id == topic.id